raise NotImplementedError if `modal_title` has not been populated
Spent a little bit of time head scratching earlier today chasing down why I was receiving `KeyError None` on `self.process`. Worked through it with Tom and thought this would be a good addition.

Tested by running the examples and clicking through majority of the modals to verify that they still worked.